requirements.txt
# === CORE DEPENDENCIES ===
python-dotenv>=1.0.0

# === WEB FRAMEWORK (FastAPI) ===
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# === DATA ANALYSIS ===
pandas>=2.0.0
numpy>=1.25.0

# === VISUALIZATION ===
matplotlib>=3.7.0
plotly>=5.18.0
seaborn>=0.13.0
wordcloud>=1.9.0

# === AI AGENT FRAMEWORK ===
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-experimental>=0.0.47
langgraph>=0.0.20

# === RAG & VECTOR DB ===
chromadb>=0.4.0
langchain-chroma>=0.1.0
sentence-transformers>=2.2.2
langchain-huggingface>=0.0.1

# === KNOWLEDGE GRAPH RAG (Optional) ===
# lightrag-hku>=0.0.8  # Disabled - API unstable

# === NLP ===
nltk>=3.8.1
textblob>=0.17.1
vaderSentiment>=3.3.2

# === ML Models ===
scikit-learn>=1.3.0

# === DEEP LEARNING (Optional) ===
torch>=2.1.0
transformers>=4.30.0

# === WEB UI ===
streamlit>=1.28.0

# === UTILITIES ===
openpyxl>=3.1.0
